Refactor Signup Page into components
